글 등록 후 상세화면으로 이동 구현

+++ src/main/java/kr/or/ddit/board/dao/PostDaoImpl.java
@@ -12,7 +12,7 @@ public class PostDaoImpl implements IPostDao{
 	
 	@Override
 	public int insertPost(SqlSession sqlSession, PostVo postVo) {
-		int insertCnt = sqlSession.insert("post.insertPost", postVo);
+		int insertCnt = (int)sqlSession.insert("post.insertPost", postVo);
 		return insertCnt;
 	}
